refactor: delete comments and unnecessary files
