+++ src/pages/GraphQl/GraphQl.tsx
@@ -1,13 +1,11 @@
 import { Documentation } from '../../components/Documentation/Documentation';
-// import { QueryEditor } from '../../components/Editor/QueryEditor';
-import { QueryEditorTextarea } from '../../components/Editor/QueryEditorTextarea';
+import { QueryEditor } from '../../components/Editor/QueryEditor';
 
 export const GraphQl: React.FC = () => {
   return (
     <>
       <Documentation />
-      {/* <QueryEditor /> */}
-      <QueryEditorTextarea />
+      <QueryEditor mode="editor" />
     </>
   );
 };
