+++ src/utils/queryEditor/index.ts
@@ -1,0 +1,4 @@
+import { formatter } from './formatter';
+import { getTabCount } from './getTabCount';
+
+export { formatter, getTabCount };
